requirements.txt
# Core AI Framework
langchain
langchain-groq

# Environment & CLI
python-dotenv
click

tree-sitter==0.21.0
tree-sitter-languages==1.10.2
pydantic 
